logo, clock, ti, outlines, and diff shadow style.
I may get rid of clock, it depends on how useful it turns out to be. ti
was a pretty good idea though.

The shadow style looks a bit weird, but I will need to do something like
that if I want to separate the head and body (which is still debatable I
guess.

I enabled outlines for sprites larger than 1 tile. I tried just drawing
each tile, but for just 5 houses, my cpu went up by about 12%. This way
is much more efficient (and less code) than drawing each tile outline
individually.

I also changed sprites/map in this commit. The lost woods was blocked on
the map, I fixed that. I deleted some of the houses in the map (so you
can go behind a house). I removed the outline for every object/actor
sprite as well. They are all drawn by outline now.

I may have edited the music also (a drum beat). And simplify the logo
animation. Not sure if that was this commit or the prev commit though.