Changed sprites (for t/b)
Removed dark lank (he won't be a character anymore, if he is, we can
just draw a shadow). Top row of characters is reserved for lank. Made
lank legs and body. Removed empty life (using health bar now). Changed
the 'X' for items to something that looks better with outlines (may
change again in the future).