Created trigger object. Used trigger on house.
Hopefully the trigger is efficient enough. I could make it more
efficient by only collision with the player. I'll do that if it becomes
a problem. I can use the trigger for other things too.

Should the trigger be only one time? No, it shouldn't because it is
reused for both events and npc/sign/house.