brang, bomb, api changes
The boomerang and bomb are both slightly better. The boomerang gives
more control, and the bomb at least blows up with no animation now.

I also made a helper function (too many tokens) of a button helper. It
returns a number based on the direction (-1, 0, 1), similar to the sgn
function I guess.

It might be interesting to return a number that gets bigger (how many
seconds you have held the button down), then the sgn function could
convert it to a number. But I don't see how I would use that at the
moment, so screw that idea.

And I added functionality for a tl to not loop. It is intertwined with
actors, so it would be nice to un-intertwine that. Oh, I have an idea!
(next commit :).