Working on standardizing a massive view object.
The reason is so that I can have multiple screens. This is still in
progress. I am getting rid of the cur_room variable. So the current room
information will go straight into the view object.

The goal is for the view object to be like the camera object. It
probably wouldn't be possible/feasible to pass everywhere.