Tbox and quotes commit.
Tbox is now powered by gun_vals. Need to update this for the sign and
lark though. Also, single quotes in code now means that the value inside
could be tokenized (change to short variable name). Double quotes means
that everything within the double quotes will not be touched. Note that
spaces with single quotes are also untouched.

Also, something like this: '"'..'"' is not handled, because the perl
parser only sees the two double quotes and hides off the code. Make sure
you only have 2 double quotes on a line if you really mean it.

I also made it so that the text case inside double quotes is inverted no
matter what. This is good for the stylistic approach I think I will take
with Zeldo. There shouldn't be any problems with this too.