Created files for a mapbuilder.
I think the next step towards developing this game is to have a better
map builder. Right now, this can only be done by manually changing the
code, that is hard to visualize though and can only be done by me.

My idea for the map builder is that you run it through a bash script. It
executes the map builder. When you click save on the map builder, it
prints the result (pipe that to a file), and exits pico-8. Using that
workflow, it will be easy to share the data from the map builder with
the game.

I was thinking of giving the map builder basic functionality. You'll be
able to pick which thing you want to make, place it & move it. You can
only place things on a grid basis of 4 pixels. So it locks at .5 the
8 pixel width.

Supported operations will be:
- Create/Place things
- Move things
- Delete things
- Save everything
- Switch to different screens
- Change the music.

Operations I don't think I need to support are:
- Creating new rooms.
- Resizing rooms.
- Configuring object parameters (But maybe....).