[GH-9] Removed double quotes. Bug fix. Compile switch between prod & dev.
The case is for removing the double quotes, but there was a slight bug
from the previous case. The bug was an off by one error, because the
program was loading `[0 - #x)`, but reading `[1 - #x]`. I fixed the bug
while keeping the token count static by explicitly specifying 0x4300.

I also changed the dev/prod behaviour. It's more expensive to build the
prod game, so the compile and watcher commands will run dev by default.
If you want it to run prod instead, specify a `-p`. This also means I
can use the same code filename and the `zeldo.p8` file doesn't have to
be changed anymore.