[GH-24] 7430. Separate debug mode from saving.
Debug mode will not listen to save state. And non debug mode only
listens to save state. The token count went down, because I removed some
code I was using to test.

Also, the init function has no save state. Once game_init is called,
that's when it loads the data now. This is a minor detail that will
probably make the game look better in the future.