[GH-41] 7349 Removed two unused lines.
The main purpose of this commit was to bring the token count below
anything I've every recorded. The lowest count I recorded in all the git
log history was 7355 with commit hash 6bbcf42 on 2021-01-09.

There is even more functionality and bug fixes to the game since that
commit a week ago. That commit was also about 48 commits ago! And the
token count has gone down. So, good job me :smile:.